Removed the settings and requirements folders
This commit removes the settings architecture and the requirements
architecture.

It simplifies the settings.py. If a user has a file within the
`alliance/config/` called `local_settings.py`, it will be imported and
will override any settings in the `settings.py` file.

Changed the namespacing across a few files.